TrafficLight: enabled property, reservation
`TrafficLight` now has an enabled property, which is always true in the
base class but delegates to `Device#isConnected` in the
`PhysicalTrafficLight`, and is also re-implemented in
multi-traffic-light.js

The reservation (check-out/check-in) concept was moved from `Device` to
`TrafficLight`, because it's more general and should be at the base of
the domain model. A "Reservable" concept (interface) could be conceived.

`Device#trafficLight` changed from a method to a property.

The concept of "available" was also moved and is not part of the
`DeviceManager` anymore.