Make usb-detection an optional dependency
The 'usb-detection' dependency is not working with preinstall for some
combination of versions, and needs to be built locally. So it's now an optional dependency.

To skip the dependency, use the `--no-optional` argument to `npm install`.

Of course, without `usb-detection`, devices will not be detected in
real-time, so it's still recommended to build it locally.

Fixes #1